daemon: implement razercommon from driver and get_serial in userspace
Instead of having to maintain a kernel driver and going through the pain
of dkms, the Razer devices could make use of hidraw directly and thus
rely only on a userspace daemon.

Tested only on the Imperator 2011, thus activated only for it.

Devices objects can opt-in to directly use the hidraw interface by
overriding the property use_hidraw.